Further card payment refactoring
Can now recover from vendor having a profile ID when we don't.
The card functions to check validation and luhn number as well as the card types enumeration moved to the card class as static/consts
Reworked table to allow a vendor ID and added some more indexes
CustomerProfile split into a 'base' version and extended version to allow for the base one to be used during testing.
Vendor refused transactions close in the DB and get labelled correctly
Added return types to TransactionManager functions