Continued SupportTicket Work
Split listing serialization to allow for some properties to be different between user and agent.
Changed category to categoryCode in a lot of places to clarify what it is.
Another pass at category descriptions.
Added 'type' filter to view (unworking presently) and rearranged controls; jump to is now separate to make it clear it's not a filter.